Add prepare_command script to podspec to build hterm if needed

+++ bin/build_hterm.sh
@@ -6,6 +6,12 @@ HTERM_VERSION='1.89'
 
 CURRENT_DIR=$(cd "$(dirname "$0")" || exit 1; pwd)
 PROJECT_DIR=$(cd "${CURRENT_DIR}/.." || exit 1; pwd)
+
+if [ -s "${CURRENT_DIR}/Terminal/Resources/hterm_all.js" ]; 
+	then echo "hterm_all.js exists. skip building else"
+	exit 0
+fi
+
 TMPDIR=$(mktemp -d)
 trap 'rm -rf "${TMPDIR}"' EXIT
